[mypyc] Refactor: extract environment class and generator related code
